rresolve merge

+++ backend/src/main/java/com/example/social_media/repository/NotificationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface NotificationRepository extends JpaRepository<Notification, Long> {
+public interface NotificationRepository extends JpaRepository<Notification, Integer> {
     Page<Notification> findByUserId(Integer userId, Pageable pageable);
     Optional<Notification> findByIdAndUserId(Integer id, Integer userId); // Thêm để kiểm tra quyền
 }
